foodgram/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-filter==22.1
django-import-export==2.8.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.4.0
et-xmlfile==1.1.0
flake8==4.0.1
idna==3.3
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
